resolved problems with data typa that i changed early

+++ src/main/java/ua/org/kpik/simple_timetable_generator/service/ExcelService.java
@@ -48,7 +48,7 @@ public class ExcelService {
                 if (teacherName == null || teacherName.trim().isEmpty() || teacherName.contains("Всього годин")) {
                     continue;
                 }
-                int hours = (int) row.getCell(3).getNumericCellValue();
+                double hours =  row.getCell(3).getNumericCellValue();
 
                 Group group = groupRepository.findByGroupName(groupName)
                         .orElseGet(() -> {
